Massive set of changes to appease MSVC++ A lot of #include "StdAfx.h" in both projects' cpp files Dumped "shawn"'s GL2API.{h|cpp} in favor of GLee Replaced (all?) remaining sprintf calls with stringstreams and the << operator Fixed backwards depth sorting

+++ CibraryEngine/BinaryChunk.cpp
@@ -1,3 +1,4 @@
+#include "StdAfx.h"
 #include "BinaryChunk.h"
 
 #include "Serialize.h"
@@ -46,7 +47,7 @@ namespace CibraryEngine
 	void ChunkTypeIndexer::HandleChunk(BinaryChunk& chunk)
 	{
 		istringstream ss(chunk.data);
-		while(ss.tellg() < chunk.data.length())
+		while(ss.tellg() < (signed int)chunk.data.length())
 		{
 			BinaryChunk part;
 			part.Read(ss);
