+++ CibraryEngine/MenuScreen.cpp
@@ -1,3 +1,4 @@
+#include "StdAfx.h"
 #include "MenuScreen.h"
 
 #include "ProgramWindow.h"
@@ -72,7 +73,7 @@ namespace CibraryEngine
 			if(window->IsActive())
 			{
 				glColor4f(1.0f, 1.0f, 1.0f, 1.0f);
-				cursor->Draw(input_state->mx, input_state->my);
+				cursor->Draw((float)input_state->mx, (float)input_state->my);
 			}
 
 			GLDEBUG();
