+++ CibraryEngine/TimingInfo.h
@@ -9,13 +9,13 @@ namespace CibraryEngine
 	{
 		public:
 			/** How much time has elapsed since the last tick */
-			double elapsed;
+			float elapsed;
 			/** The total amount of time that has passed in the game */
-			double total;
+			float total;
 
 			/** Initializes a TimingInfo with elapsed and total both equal to zero */
 			TimingInfo();
 			/** Initializes a TimingInfo with both values specified */
-			TimingInfo(double elapsed, double total);
+			TimingInfo(float elapsed, float total);
 	};
 }
